refactor: add test config setting and modify app structure
