+++ .env.test
@@ -1,0 +1,7 @@
+ENV=testing
+POSTGRES_USER=test_user
+POSTGRES_PASSWORD=test_password
+POSTGRES_DB=test_db
+
+DB_HOST=db
+DB_PORT=5433
