+++ app/config.py
@@ -13,6 +13,16 @@ class Settings(BaseSettings):
     model_config = SettingsConfigDict(env_file=".env")
 
 
+class TestSettings(BaseSettings):
+    postgres_db: str
+    postgres_user: str
+    postgres_password: str
+    db_host: str
+    db_port: str
+
+    model_config = SettingsConfigDict(env_file=".env.test")
+
+
 # ! To be used as a dependency injection in requests
 # ? Why?
 # * to provide a different settings object during testing by creating a
@@ -23,6 +33,11 @@ class Settings(BaseSettings):
 # * Reading a file from disk is normally a costly (slow) operation so its
 # * better to do it once then
 # * reuse the same settings object, instead of reading it for each request.
-# @lru_cache
-# def get_settings():
-#     return Settings()
+@lru_cache
+def get_settings():
+    return Settings()
+
+
+@lru_cache()
+def get_test_settings():
+    return TestSettings()
